Eager to know if this will work. See #113.
